Extract shared event-form field components
